+++ admin/post-edit.php
@@ -337,7 +337,7 @@ if ($post->published_at) {
     <link rel="stylesheet" href="/admin/assets/easymde.min.css">
     <link rel="stylesheet" href="/admin/assets/font-awesome.min.css">
 </head>
-<body class="admin-page">
+<body class="admin-page" data-slug-type="post"<?= $post?->id ? ' data-slug-id="' . $post->id . '"' : '' ?>>
 
 <?php require __DIR__ . '/partials/nav.php'; ?>